refactor(driver): extract flush_to_l1 + helpers to driver/flush.rs (refactor 2.1k)
Move 5 methods from driver/mod.rs to driver/flush.rs:
- flush_to_l1: main L1 submission pipeline (FlushPlan construction,
  hold arming, receipt handling, rewind on mismatch)
- verify_trigger_receipts: wait for L2→L1 trigger receipts
- send_l2_to_l1_triggers: explicit-nonce executeL2TX calls
- forward_queued_l1_txs: forward raw user L1 txs after postBatch
- compute_gas_overbid: gas price overbid from queued L1 user txs

mod.rs: 3002 → 2041 lines (-961). flush.rs: 1019 lines.
Zero behavioral change. get_l1_provider promoted to pub(super).
Unused imports cleaned (GasPriceHint, U256, SolCall removed from
mod.rs; MAX_BATCH_SIZE, SUBMISSION_COOLDOWN_SECS, TriggerExecutionResult
moved to #[cfg(test)] re-exports for driver_tests.rs).

608/608 tests pass, clippy clean.
